Move all includes of GLEW to internal classes only

+++ RTICreator/MainMenuRoom.hpp
@@ -1,4 +1,4 @@
-﻿/*! 
+﻿/*!
  *  @author    Dubsky Tomas
  */
 #pragma once
@@ -9,9 +9,9 @@
 #include <RealEngine/rendering/output/Viewport.hpp>
 #include <RealEngine/rendering/textures/Texture.hpp>
 
-/**
- * @brief Room with the UI
-*/
+ /**
+  * @brief Room with the UI
+ */
 class MainMenuRoom : public RE::Room {
 public:
 	MainMenuRoom(RE::CommandLineArguments args);
@@ -39,7 +39,7 @@ private:
 
 	//View
 	RE::View2D m_texView;
-	RE::TypedBuffer m_texViewUBO{ RE::UNIF_BUF_VIEWPORT_MATRIX, sizeof(glm::mat4), RE::BufferUsageFlags::DYNAMIC_STORAGE };
+	RE::TypedBuffer m_texViewUBO{RE::UNIF_BUF_VIEWPORT_MATRIX, static_cast<int>(sizeof(glm::mat4)), RE::BufferUsageFlags::DYNAMIC_STORAGE};
 	glm::vec2 m_overlap = glm::vec2(0.2f, 0.2f);
 	glm::vec3 m_backgroundColor = glm::vec3(0.1f, 0.1f, 0.1f);
 
@@ -49,9 +49,9 @@ private:
 	size_t m_wrapStyleX = 0u;
 	size_t m_wrapStyleY = 0u;
 	glm::ivec2 m_subimagesSprites{};
-	glm::vec2 m_subimageDims{ 1.0f };
+	glm::vec2 m_subimageDims{1.0f};
 	glm::vec2 m_pivot{};
-	glm::vec4 m_borderColor{ 0.0f, 0.0f, 0.0f, 1.0f };
+	glm::vec4 m_borderColor{0.0f, 0.0f, 0.0f, 1.0f};
 
 
 	glm::vec2 m_offset = glm::vec2(0.0f, 0.0f);
